+++ src/rebalance/__init__.py
@@ -1,4 +1,4 @@
 """rebalance package."""
 
 __all__ = ["__version__"]
-__version__ = "0.5.5"
+__version__ = "0.7.0"